api_server/requirements.txt
django
django-rest-framework
django-haystack==2.6.0
wheel
mysqlclient==1.3.12
pysolr==3.7.0
